requirements.txt
flask>=3.0.0
anthropic>=0.25.0
python-nmap>=0.7.1
scapy>=2.5.0
requests>=2.31.0
reportlab>=4.1.0
python-dotenv>=1.0.0
networkx>=3.2.0
mitreattack-python>=2.0.5
cvss>=2.6
xmltodict>=0.13.0
werkzeug>=3.0.0
